feat: Some more work on the config parser
This commit creates a new parser based on a new language which hopefully
also has multi-line array support.
If you're wondering why this commit is weird then it's because I am
rebasing and squashing a lot.

Original commit message (or partially original):

This does look a tad bit ugly but that's what I get for implementing a custom config language with complex features.

Irregardless, it runs quite fast and I do not mind the "ugly" code, at least I can understand it.

In short, I am quite happy with my progress and with developing in Nim.

I am excited for what the future holds.