fix: Switching to parsecfg to fix a bug related to GCed config tables.
So instead of making my own parser, I have decided to use Nim's parsecfg module.

Also good news, the bug is gone. But I did have to use a workaround, in short, if it appears again then we're fucked.

But everything seems fine for now, turning off threadAnalysis got the thing to compile. But just because it compiles, doesn't mean it's working, but really in this case it is working.

I think everything will be fine after all.

"The good news is that when the code compiles with `.gcsafe`, it really is GC
safe and nothing to worry about."
As said by Araq himself! If Araq isn't worried then neither am I.

I am worried at the back of my mind that some mysterious thread-related bug *will* appear and force me to rewrite everything.

Signed-off-by: Leo Gavilieau <xmoo@privacyrequired.com>