docs: Taking a small break! Stay safe!
I have been working non-stop for a week which led to some amazing progress but honestly I am so overworked, and I cannot deal with all this stress.

So I am taking a small break for my own sake.
I will probably be coding a little bit here and there but most of that spontaneous effort will be placed into having a break and getting new hobbies.

Thank you all!

Signed-off-by: Leo Gavilieau <xmoo@privacyrequired.com>