feat!: Remove "Unknown" caller possibility.
in lib.nim, there is a function called error() this is called when the programs has encountered a error it cannot solve and has to exit.

By default we want to know what happened and what function called the error() function in the first place.

Up until this point, you could call error() without specifying the function name you are calling from.

Well now this is no longer valid.

BREAKING CHANGES: If your code hasn't yet adapted to this then I dont know what to tell you. Fix it, or else.

Signed-off-by: Leo Gavilieau <xmoo@privacyrequired.com>