feat: Implemented a faster replace() function.
strutils.replace() searches the entire string but searching the string over and over to replace one thing is slow.

This implementation searches until it has found a match and then replaces it. And skips the rest.

i am pretty sure strutils has something like this already, it's probably called replaceFirst() or smth.

We will use that then but this works (i think? I haven't tested it yet)

Signed-off-by: Leo Gavilieau <xmoo@privacyrequired.com>