optimize: deprecate int64ToString
It was a nice 12 commits when you stayed with us but you are a bit weird so you have to go.

This commit fixes the root problem of int64ToString.
A new function named "int64ToInt" converts int64s to ints just for the db parts that need it

And it makes booleans use parseBool for their work (which is much faster) than what we currentlyt have

Signed-off-by: Leo Gavilieau <xmoo@privacyrequired.com>