feat!: add UserType attribute to the User object
BREAKING CHANGES: Make sure your code is well-adapted to the introduction of the UserType object and the User.kind object field.
If your code loops over the fields of the User object then make sure to account for this change.
If your code just outputs existing object fields or it does some trivial action (newUser(), escape(), addUser()) then it shouldn't be affected.

Signed-off-by: Leo Gavilieau <xmoo@privacyrequired.com>