fix: Replace db.isNil() with a working boolean check for database access
We sadly can't do anything to the DbConn object without storage access errors, and so this is the stupidest, but also simplest check ever that works!

This also fixes an issue that I just discovered where pothole creates too many connections to the database, and it fails. (I think it's roughly after 100 connections? thats what max_connections seems to be set to anyway)