feat: Add "KDF" enum for handling different KDF functions.
"KDF function" is redundant, oh whatever.

This commit adds a new "KDF" object to the crypto.nim module with the sole purpose of abstracting away which KDF (Key-derivation-function, aka. Password hashing) algorithm to choose.

The main reason I am doing this is for readability purposes, you wouldn't know what "1" means as a KDF function *without* looking it up in a separate table, but if it is PBKDF_HMAC_SHA512 then you have a better idea, and also, you have a function that will automatically hash with the correct algorithm without you even knowing!

Oh and this commit also moves lib.kdf to crypto.kdf since that's more appropriate, though there is an alias provided in lib.nim for your convenience and with a deprecated marker so you know where it is!

One more object couldn't possibly hurt, right?