fix: Fix memory leak
So, without this flag, when Pothole is given a tremendous number of requests, it will eventually use all memory, and it won't give it back.

Big thanks to Phil from the Nim discord server for helping me out! Because it turns out, just adding -d:useMalloc completely fixes this problem!

As phil said, it might be because the Nim alloactor is not as reliable as malloc, or it might be because I am doing something wrong and it's completely my fault. But either way, without spending many hours, diving deep into the allocator, I won't be able to figure out whose fault it is.

And, I just can't honestly justify giving up that time when -d:useMalloc works ridiculously well with only a small slowdown. So, let's just stuff -d:useMalloc into the nimble file and call it a day.

TODO though, for whoever wants to: Figure out what the hell is happening without -d:useMalloc :P