fix: Add check for users that already exist.
This is extremely basic, and I should have done it but I missed it.

Either way though, the block below, with its catch prevents anything truly bad from happening. The database layer already prevents it by having adding "UNIQUE" to the schema.

This check is more for user friendliness reasons.