fix: Fix the function name collusion issue by nuking the dev subsystem
the db subsystem pretty much already does everything we want. So, let's use it instead.

This isn't a proper fix though, because we have more commands like user new, and post new. And if cligen can't handle these for us then this whole migration is screwed and we gotta go back to the old method. (Or use another cli helper tool)