feat: Implemented sendEmail() and sendEmailCode()
These do block the thread, but eh, it's not that big of a deal given that Mummy will use other threads to serve other users in the meantime.

it would be a big deal if a huge number of new users were created. Then there would be slow downs, but hm... I can't think of a good way to avoid this beyond just y'know disabling user registrations? or requiring approval?