fix: Add port in url and disable email verifications by default.
This config file is meant for testers, developers and so on. Users should edit it to suit their own needs.

TODO: In the future, it might be better to offer 2 config files, one (used by default) that is developer-oriented and the other with production settings meant for users to copy easily.