docs: Add note to README about prebuilt binaries being available
PS: The build automation was originally going to be powered by bocchi, a "CI" system that I never finished and that I poured several days of effort into.

Instead of bocchi, I just programmed a simple shell script to do all the building. And I've setup a cron to check for new commits daily...

Simplicity really is amazing, isn't it?
(I might put the script on a gist or in the ftp site somewhere with setup instructions... but it does not deserve its own repo)