feat: Here are the new content to html procs
I am trying tor reduce quark's API size as much as possible, just to make writing tests a bit easier.

This commit also has the benefit of reducing the amount of time dbPool is being borrowed from. Let me explain.

The old proc's used to not just fetch content from the database, but it also processed it in the same procedure.

but by splitting them and re-adjusting the code, we can make sure that only the actual database-related code has to borrow a database connection.

And that's the generally sane thing to do.

This is such a small difference, and it's not really a big deal. But the cleaner API is worth it.